구글맵 위젯 분리 및 하단 ui 위젯 분리
